ActionMapからキャラ名によるHashMapをなくす直前 敵にもActionsを実装途中

+++ Projecto01forProgress/src/elements/Element.java
@@ -38,7 +38,7 @@ public class Element {
 		this.stage = stage;
 	}
 	
-	public void moveTo(int x, int y){
+	public void moveTo(double x, double y){
 		this.x = x;
 		this.y = y;
 	}
